utilccl: split out pure license check def+logic
this is just a code-moving refactor, splitting the license def (proto) and the
pure helpers that go with it into their own package, specifically not including
the cluster settings handling parts.

Settings code, and the dependencies it pulls in are not ideal for consumption by
downstream dependees (from a purist perspective, but also a practical one, since
some of the transitive deps have init hooks).
